[BACKLOG-17024] Rename pentaho.visual.models.MetricDotAbstract to pentaho.visual.models.MetricPointAbstract.
