+++ impl/client/src/main/javascript/web/pentaho/ccc/visual/bubble.js
@@ -16,7 +16,7 @@
 define([
   "module",
   "pentaho/visual/models/bubble",
-  "./metricDotAbstract"
+  "./metricPointAbstract"
 ], function(module, modelFactory, baseViewFactory) {
 
   "use strict";
