+++ impl/client/src/main/javascript/web/pentaho/visual/models/metricPointAbstract.js
